Simplify errors framework
