--- packages/backend-core/src/errors/base.ts
@@ -1,10 +1,0 @@
-export class BudibaseError extends Error {
-  code: string
-  type: string
-
-  constructor(message: string, code: string, type: string) {
-    super(message)
-    this.code = code
-    this.type = type
-  }
-}
