--- packages/backend-core/src/errors/generic.ts
@@ -1,7 +1,0 @@
-import { BudibaseError } from "./base"
-
-export class GenericError extends BudibaseError {
-  constructor(message: string, code: string, type: string) {
-    super(message, code, type ? type : "generic")
-  }
-}
