+++ packages/backend-core/src/errors/index.ts
@@ -1,3 +1,1 @@
 export * from "./errors"
-export { UsageLimitError, FeatureDisabledError } from "./licensing"
-export { HTTPError } from "./http"
